Video sources: rename tumlive/ module to videosource/
