+++ src/main/webapp/app/lecture/manage/lecture-units/services/attachment-video-unit.service.ts
@@ -106,7 +106,7 @@ export class AttachmentVideoUnitService {
      */
     getPlaylistUrl(pageUrl: string): Observable<string | undefined> {
         const params = new HttpParams().set('url', pageUrl);
-        return this.httpClient.get('/api/tumlive/playlist', { params, responseType: 'text' }).pipe(catchError(() => of(undefined)));
+        return this.httpClient.get('/api/videosource/playlist', { params, responseType: 'text' }).pipe(catchError(() => of(undefined)));
     }
 
     /**
