--- src/test/java/de/tum/cit/aet/artemis/tumlive/architecture/TumLiveCodeStyleArchitectureTest.java
@@ -1,11 +1,0 @@
-package de.tum.cit.aet.artemis.tumlive.architecture;
-
-import de.tum.cit.aet.artemis.shared.architecture.module.AbstractModuleCodeStyleTest;
-
-class TumLiveCodeStyleArchitectureTest extends AbstractModuleCodeStyleTest {
-
-    @Override
-    public String getModulePackage() {
-        return ARTEMIS_PACKAGE + ".tumlive";
-    }
-}
