--- src/test/java/de/tum/cit/aet/artemis/tumlive/architecture/TumLiveResourceArchitectureTest.java
@@ -1,10 +1,0 @@
-package de.tum.cit.aet.artemis.tumlive.architecture;
-
-import de.tum.cit.aet.artemis.shared.architecture.module.AbstractModuleResourceArchitectureTest;
-
-class TumLiveResourceArchitectureTest extends AbstractModuleResourceArchitectureTest {
-
-    public String getModulePackage() {
-        return ARTEMIS_PACKAGE + ".tumlive";
-    }
-}
