+++ src/test/java/de/tum/cit/aet/artemis/videosource/architecture/VideoSourceCodeStyleArchitectureTest.java
@@ -1,0 +1,11 @@
+package de.tum.cit.aet.artemis.videosource.architecture;
+
+import de.tum.cit.aet.artemis.shared.architecture.module.AbstractModuleCodeStyleTest;
+
+class VideoSourceCodeStyleArchitectureTest extends AbstractModuleCodeStyleTest {
+
+    @Override
+    public String getModulePackage() {
+        return ARTEMIS_PACKAGE + ".videosource";
+    }
+}
