+++ src/test/java/de/tum/cit/aet/artemis/videosource/architecture/VideoSourceRepositoryArchitectureTest.java
@@ -1,0 +1,10 @@
+package de.tum.cit.aet.artemis.videosource.architecture;
+
+import de.tum.cit.aet.artemis.shared.architecture.module.AbstractModuleRepositoryArchitectureTest;
+
+class VideoSourceRepositoryArchitectureTest extends AbstractModuleRepositoryArchitectureTest {
+
+    public String getModulePackage() {
+        return ARTEMIS_PACKAGE + ".videosource";
+    }
+}
