+++ src/test/java/de/tum/cit/aet/artemis/videosource/architecture/VideoSourceServiceArchitectureTest.java
@@ -1,0 +1,10 @@
+package de.tum.cit.aet.artemis.videosource.architecture;
+
+import de.tum.cit.aet.artemis.shared.architecture.module.AbstractModuleServiceArchitectureTest;
+
+class VideoSourceServiceArchitectureTest extends AbstractModuleServiceArchitectureTest {
+
+    public String getModulePackage() {
+        return ARTEMIS_PACKAGE + ".videosource";
+    }
+}
